remove dataio
