+++ pkg/vm/engine/aoe/storage/buffer/node/cleaner.go
@@ -15,22 +15,20 @@
 package node
 
 import (
-	ioif "matrixone/pkg/vm/engine/aoe/storage/dataio/iface"
+	"matrixone/pkg/vm/engine/aoe/storage/buffer/node/iface"
 	"os"
-	// log "github.com/sirupsen/logrus"
 )
 
 type NodeCleaner struct {
 	Filename []byte
 }
 
-func NewNodeCleaner(filename []byte) ioif.Cleaner {
+func NewNodeCleaner(filename []byte) iface.Cleaner {
 	return &NodeCleaner{
 		Filename: filename,
 	}
 }
 
 func (nc *NodeCleaner) Clean() error {
-	// log.Infof("NodeCleaner removing %s", nc.Filename)
 	return os.Remove(string(nc.Filename))
 }
