+++ pkg/vm/engine/aoe/storage/buffer/node/iface/io.go
@@ -1,0 +1,25 @@
+package iface
+
+type Reader interface {
+	Load() error
+}
+
+type Writer interface {
+	Flush() error
+}
+
+type Cleaner interface {
+	Clean() error
+}
+
+type IOHandle interface {
+	Reader
+	Writer
+	Cleaner
+}
+
+type DefaultIOHandle struct {
+	Reader
+	Writer
+	Cleaner
+}
