+++ pkg/vm/engine/aoe/storage/buffer/node/types.go
@@ -17,16 +17,14 @@ package node
 import (
 	buf "matrixone/pkg/vm/engine/aoe/storage/buffer"
 	mgrif "matrixone/pkg/vm/engine/aoe/storage/buffer/manager/iface"
-	nif "matrixone/pkg/vm/engine/aoe/storage/buffer/node/iface"
+	"matrixone/pkg/vm/engine/aoe/storage/buffer/node/iface"
 	"matrixone/pkg/vm/engine/aoe/storage/common"
-	ioif "matrixone/pkg/vm/engine/aoe/storage/dataio/iface"
 	"sync"
 )
 
 type NodeBuffer struct {
 	buf.IBuffer
 	ID uint64
-	// Type nif.BufferType
 }
 
 type NodeHandleCtx struct {
@@ -42,17 +40,17 @@ type NodeHandleCtx struct {
 
 type NodeHandle struct {
 	sync.Mutex
-	State       nif.NodeState
+	State       iface.NodeState
 	ID          uint64
 	Buff        buf.IBuffer
 	Spillable   bool
 	File        common.IVFile
 	UseCompress bool
-	RTState     nif.NodeRTState
+	RTState     iface.NodeRTState
 	Refs        uint64
 	Manager     mgrif.IBufferManager
 	Iter        uint64
-	IO          ioif.IO
+	IO          iface.IOHandle
 	Constructor buf.MemoryNodeConstructor
 }
 
@@ -60,6 +58,6 @@ type NodeHandle struct {
 // The following IBufferManager::Pin will call INodeHandle::Ref to increment the reference count
 // BufferHandle should alway be closed manually when it is not needed, which will call IBufferManager::Unpin
 type BufferHandle struct {
-	Handle  nif.INodeHandle
+	Handle  iface.INodeHandle
 	Manager mgrif.IBufferManager
 }
